Override ogent code to always return ent errors.
